+++ src/DependencyInjection/SentryExtension.php
@@ -4,6 +4,7 @@ namespace Sentry\SentryBundle\DependencyInjection;
 
 use Monolog\Logger as MonologLogger;
 use Sentry\ClientBuilderInterface;
+use Sentry\Integration\IgnoreErrorsIntegration;
 use Sentry\Monolog\Handler;
 use Sentry\Options;
 use Sentry\SentryBundle\ErrorTypesParser;
@@ -11,6 +12,7 @@ use Sentry\SentryBundle\EventListener\ErrorListener;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Exception\LogicException;
 use Symfony\Component\DependencyInjection\Loader;
 use Symfony\Component\DependencyInjection\Reference;
@@ -63,7 +65,6 @@ class SentryExtension extends Extension
             'default_integrations',
             'enable_compression',
             'environment',
-            'excluded_exceptions',
             'http_proxy',
             'logger',
             'max_request_body_size',
@@ -90,6 +91,10 @@ class SentryExtension extends Extension
             $options->addMethodCall('setInAppExcludedPaths', [$processedOptions['in_app_exclude']]);
         }
 
+        if (\array_key_exists('in_app_include', $processedOptions)) {
+            $options->addMethodCall('setInAppIncludedPaths', [$processedOptions['in_app_include']]);
+        }
+
         if (\array_key_exists('error_types', $processedOptions)) {
             $parsedValue = (new ErrorTypesParser($processedOptions['error_types']))->parse();
             $options->addMethodCall('setErrorTypes', [$parsedValue]);
@@ -114,14 +119,25 @@ class SentryExtension extends Extension
             $options->addMethodCall('setClassSerializers', [$classSerializers]);
         }
 
+        $integrations = [];
         if (\array_key_exists('integrations', $processedOptions)) {
-            $integrations = [];
             foreach ($processedOptions['integrations'] as $integrationName) {
                 $integrations[] = new Reference(substr($integrationName, 1));
             }
-
-            $options->addMethodCall('setIntegrations', [$integrations]);
         }
+
+        if (\array_key_exists('excluded_exceptions', $processedOptions) && $processedOptions['excluded_exceptions']) {
+            $ignoreOptions = [
+                'ignore_exceptions' => $processedOptions['excluded_exceptions'],
+            ];
+
+            $integrations[] = new Definition(IgnoreErrorsIntegration::class, [$ignoreOptions]);
+        }
+
+        $integrationsCallable = new Definition('callable', [$integrations]);
+        $integrationsCallable->setFactory([IntegrationFilterFactory::class, 'create']);
+
+        $options->addMethodCall('setIntegrations', [$integrationsCallable]);
     }
 
     private function valueToCallable($value)