+++ CHANGELOG.md
@@ -6,12 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
  - ...
+ 
+## 3.4.0 (2020-01-17)
+ - Add support for `sentry/sentry` 2.3 (#298)
+ - Drop support for `sentry/sentry` < 2.3 (#298)
+ - Add support to `in_app_include` client option (#298)
+ - Remap `excluded_exception` option to use the new `IgnoreErrorIntegration` (#298)
 
 ## 3.3.2 (2020-01-16)
-- Fix issue with exception listener under Symfony 4.3 (#301)
+ - Fix issue with exception listener under Symfony 4.3 (#301)
 
 ## 3.3.1 (2020-01-14)
-- Fixed Release
+ - Fixed Release
 
 ## 3.3.0 (2020-01-14)
  - Add support for Symfony 5.0 (#266, thanks to @Big-Shark)